refactor: 831 - change file names to kebabcase style and add linter rule
