+++ dev-data/aws-fundamentals.data.ts
