+++ dev-data/communication.data.ts
@@ -1,4 +1,4 @@
-import { DISCORD_LINKS } from './courseTitles.data';
+import { DISCORD_LINKS } from './course-titles.data';
 
 export const RS_DOCS_COMMUNICATION_LINK = 'https://docs.rs.school/#/rs-school-chats';
 export const RS_DOCS_EN_LINK = 'https://docs.rs.school/#/en/';
