+++ dev-data/mentorship-data.types.ts
@@ -1,6 +1,6 @@
 import { StaticImageData } from 'next/image';
 
-import { COURSE_TITLES } from './courseTitles.data';
+import { COURSE_TITLES } from './course-titles.data';
 import { ROUTES } from '@/core/const';
 import { Language } from '@/shared/types';
 import { SocialMediaProps } from '@/shared/ui/social-media-item';
