+++ dev-data/mentorship.data.ts
@@ -1,4 +1,4 @@
-import { COURSE_TITLES } from './courseTitles.data';
+import { COURSE_TITLES } from './course-titles.data';
 import { ROUTES } from '@/core/const';
 import angular from '@/shared/assets/icons/angular.svg';
 import javascript from '@/shared/assets/icons/javascript.svg';
@@ -274,7 +274,8 @@ export const mentorshipCourses: MentorshipCourse[] = [
       {
         id: 1,
         title: 'Проведение технического интервью',
-        description: 'По результату которого ментор принимает решение о работе со студентом. Интервью проводятся с 2+ студентами - общий размер команды определяется ментором в зависимости от его загруженности',
+        description:
+          'По результату которого ментор принимает решение о работе со студентом. Интервью проводятся с 2+ студентами - общий размер команды определяется ментором в зависимости от его загруженности',
         icon: activitiesContent.icon.interview,
         links: [
           {
@@ -286,13 +287,15 @@ export const mentorshipCourses: MentorshipCourse[] = [
       {
         id: 2,
         title: 'Лучшие практики',
-        description: 'Ментор показывает, какие конструкции кода лучше всего использовать, основываясь на своем опыте. Подсказывает как лучше организовать структуру приложения',
+        description:
+          'Ментор показывает, какие конструкции кода лучше всего использовать, основываясь на своем опыте. Подсказывает как лучше организовать структуру приложения',
         icon: activitiesContent.icon.bestPracticesJS,
       },
       {
         id: 3,
         title: 'Ревью кода',
-        description: 'Ментор просматривает код, указывает на неточности, предлагает, как их улучшить. Ревью кода помогает студенту развить необходимые навыки для работы в команде',
+        description:
+          'Ментор просматривает код, указывает на неточности, предлагает, как их улучшить. Ревью кода помогает студенту развить необходимые навыки для работы в команде',
         icon: activitiesContent.icon.codeReview,
         links: [
           {
@@ -304,13 +307,15 @@ export const mentorshipCourses: MentorshipCourse[] = [
       {
         id: 4,
         title: 'Проведение двух CoreJS интервью',
-        description: 'Ментор проводит два интервью со студентами других менторов по основным темам курса. Участие в интервью позволяет студентам не только увидеть свои пробелы в знаниях, но и почувствовать атмосферу подобных встреч',
+        description:
+          'Ментор проводит два интервью со студентами других менторов по основным темам курса. Участие в интервью позволяет студентам не только увидеть свои пробелы в знаниях, но и почувствовать атмосферу подобных встреч',
         icon: activitiesContent.icon.interviewCore,
       },
       {
         id: 5,
         title: 'Курирование командного задания',
-        description: 'Опыт и практические знания ментора помогут организовать студентов для работы над командным проектом',
+        description:
+          'Опыт и практические знания ментора помогут организовать студентов для работы над командным проектом',
         icon: activitiesContent.icon.finalProject,
       },
     ],
