+++ package.json
@@ -73,6 +73,7 @@
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "eslint-plugin-sort-exports": "^0.9.1",
+    "eslint-plugin-unicorn": "^58.0.0",
     "globals": "^16.0.0",
     "husky": "^9.1.7",
     "jsdom": "^26.0.0",
