+++ src/entities/course/helpers/sync-with-api-data.ts
@@ -1,6 +1,6 @@
 import { COURSE_DATE_FORMAT, MENTORING_DATE_FORMAT } from '@/entities/course/constants';
 import { ApiCoursesResponse } from '@/entities/course/types';
-import { dayJS } from '@/shared/helpers/dayJS';
+import { dayJS } from '@/shared/helpers/day-js';
 import { courses as baseCourseData } from 'data';
 
 export const syncWithApiData = (apiCourses: ApiCoursesResponse[]) => {
