+++ src/shared/helpers/get-course-date.ts
