+++ src/views/mentorship/helpers/transform-courses-to-mentorship.ts
