+++ src/widgets/hero-course/ui/hero-course.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image';
 
 import { getCourseStatus } from '../helpers/get-course-status';
 import { LABELS } from '@/shared/constants';
-import { dayJS } from '@/shared/helpers/dayJS';
+import { dayJS } from '@/shared/helpers/day-js';
 import { selectCourse } from '@/shared/hooks/use-course-by-title/utils/select-course';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { MainTitle } from '@/shared/ui/main-title';
